requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
boto==2.49.0
boto3==1.13.18
botocore==1.16.18
bs4==0.0.1
bson==0.5.9
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
dominate==2.5.1
entrypoints==0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
flask-restplus==0.13.0
flask-swagger-ui==3.25.0
gast==0.3.3
gensim==3.8.3
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-speech==1.3.2
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
LEPL==5.1.3
llvmlite==0.34.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numba==0.51.0
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==1.9.6
Pygments==2.6.1
pylint==2.4.4
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==2.0.0
soupsieve==2.0.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
validator-collection==1.4.2
validators==0.18.1
visitor==0.1.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.11.2
zipp==3.1.0
